ECOPROJECT-4492 | feat: adding cluster recommendations info in the example report
